+++ examples/credential/main.tf
@@ -1,23 +1,25 @@
 provider "cdp" {
   # cdp_access_key_id = ""
   # cdp_private_key = ""
-  # cdp_profile = "default"
+  # cdp_profile = "dev"
+  # cdp_config_file = "/Users/spongebob/.cdp/config"
+  # cdp_shared_credentials_file = "/Users/spongebob/.cdp/credentials"
 }
 
 resource "cdp_environments_aws_credential" "my-credentials" {
-  name = "my-cloudera-cdp-credentials"
-  role_arn = "arn:aws:iam::643594445606:role/EGRESS_TEST_cdp-cross-account-role"
+  credential_name = "my-cloudera-cdp-credentials"
+  role_arn = "arn:aws:iam::111111111111:role/my-cdp-cross-account-role"
   description = "testing the terrafrom-provider-cdp"
 }
 
-output "name" {
-  value = cdp_environments_credential.my-credentials.name
+output "credential_name" {
+  value = cdp_environments_aws_credential.my-credentials.credential_name
 }
 
 output "role_arn" {
-  value = cdp_environments_credential.my-credentials.role_arn
+  value = cdp_environments_aws_credential.my-credentials.role_arn
 }
 
 output "crn" {
-  value = cdp_environments_credential.my-credentials.crn
+  value = cdp_environments_aws_credential.my-credentials.crn
 }